+++ examples/firebase/src/containers/App/index.tsx
@@ -1,9 +1,7 @@
 import { useState } from 'react'
-import { loadDeveloperTools } from '@signaldb/core'
+import '@signaldb/devtools'
 import Todos from '../../models/Todos'
 import List from './List'
-
-loadDeveloperTools()
 
 const App: React.FC = () => {
   const [text, setText] = useState('')