+++ packages/devtools/devtools/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Added logic to load developer tools to this package
+
 ## [1.0.0-beta.3] - 2025-01-23
 
 ### Changed